feat: compacted the service file and added corresponding notebook changes
